Added a clarification note to self.
Some instructions need more clock cycles in some situations.
For example, an instruction may need an extra clock cycle if it accesses an address that crosses the memory page boundaries. Other instructions might need an extra clock cycle if it uses a particular addressing mode.

If a particular instruction needs an additional clock cycle because of a slow addressing mode, it's known from the instruction format itself, thus the implementation in the `clock` function.

According to the datasheet, there're some instructions that need one instruction if the branching address is on the same memory page, i.e. anyway, and add two additional clock cycles only if the branching address is on a different memory page, i.e. crossed memory page boundaries. either to add one or two additional instruction cycles is decided during the instruction execution itself.